course2/4.algo_datastruct/4.datastruct_tree/task2.4.4.2/main.go
package main

import (
	"fmt"
	"github.com/google/btree"
)

type User struct {
	ID   int
	Name string
	Age  int
}

func (u User) Less(than btree.Item) bool {
	return u.ID < than.(User).ID
}

type BTree struct {
	tree *btree.BTree
}

func NewBTree(degree int) *BTree {
	return &BTree{
		tree: btree.New(degree),
	}
}

func (bt *BTree) Insert(user User) {
	bt.tree.ReplaceOrInsert(user)
}

func (bt *BTree) Search(id int) *User {
	user := User{ID: id}

	item := bt.tree.Get(user)
	if item != nil {
		fnUser := item.(User)
		return &fnUser
	}
	return nil
}

func main() {
	bt := NewBTree(2)
	users := []User{
		{1, "Alice", 30},
		{2, "Bob", 25},
		{3, "Charlie", 35},
		// добавьте больше пользователей при необходимости
	}

	for _, user := range users {
		bt.Insert(user)
	}
	user := bt.Search(2)
	if user != nil {
		fmt.Printf("Найден пользователь: %v\n", *user)
	} else {
		fmt.Println("Пользователь не найден")
	}
}
